Complete permission interceptor

+++ src/main/java/org/teamwe/carrent/controller/OrderController.java
@@ -2,11 +2,14 @@ package org.teamwe.carrent.controller;
 
 import org.springframework.web.bind.annotation.*;
 import org.teamwe.carrent.controller.utils.Format;
+import org.teamwe.carrent.controller.utils.SessionAttr;
 import org.teamwe.carrent.entity.Order;
+import org.teamwe.carrent.entity.User;
 import org.teamwe.carrent.service.OrderService;
 import org.teamwe.carrent.utils.ReturnStatus;
 import org.teamwe.carrent.utils.StringUtil;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 /**
@@ -48,7 +51,13 @@ public class OrderController {
     }
 
     @DeleteMapping("/order/{orderId}")
-    public Format cancelOrder(@PathVariable String orderId) {
+    public Format cancelOrder(@PathVariable String orderId, HttpSession session) {
+        Order o = service.getById(orderId);
+        if (o == null ||
+                session.getAttribute(SessionAttr.USER_ID) == null ||
+                !session.getAttribute(SessionAttr.USER_ID).equals(o.getEmail())) {
+            return new Format().code(ReturnStatus.FAILURE).message("Illegal user");
+        }
         return new Format().code(service.deleteOrder(orderId));
     }
 
@@ -56,13 +65,20 @@ public class OrderController {
     public Format endOrder(@PathVariable String id) {
         return new Format().code(service.endOrder(id));
     }
+
     @GetMapping("/order/{id}")
-    public Format getOrder(@PathVariable String id) {
+    public Format getOrder(@PathVariable String id, HttpSession session) {
         Order o = service.getById(id);
         if (o == null) {
             return new Format().code(ReturnStatus.FAILURE).message("No such of order : " + id);
-        } else {
+        }
+        String type = "" + session.getAttribute(SessionAttr.USER_TYPE);
+        String email = "" + session.getAttribute(SessionAttr.USER_ID);
+
+        if (String.valueOf(User.ENGENEER).equals(type)
+                || email.equals(o.getEmail())) {
             return new Format().code(ReturnStatus.SUCCESS).addData("order", o);
         }
+        return new Format().code(ReturnStatus.FAILURE).message("No such of order : " + id);
     }
 }
